requirements.txt
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
html2text==2020.1.16
